+++ CHANGELOG.md
@@ -145,6 +145,8 @@ CHANGELOG
   - Fix bug preventing use of old format after using the new format.
 
 #### Experts
+- `intelmq.bots.experts.cymru_whois.expert`: Add optional parameter `overwrite`, current behavior was `True`, default if not given is `True` now, will change to `False` in 3.0.0 (#1452, #1455).
+- `intelmq.bots.experts.reverse_dns.expert`: Add optional parameter `overwrite`, current behavior was `True`, default if not given is `True` now, will change to `False` in 3.0.0 (#1452, #1455).
 
 #### Outputs
 - `intelmq.bots.outputs.amqptopic.output`: use default SSL context for client purposes, fixes compatibility with python < 3.6 if TLS is used.
@@ -167,6 +169,7 @@ CHANGELOG
   - No error message for disabled bots on botnet reload.
   - Fix `upgrade-conf` is state file is empty or not existing.
   - Use arpgarse's `store_true` action for flags instead of `store_const`.
+  - If the loading of the defaults configuration failed, a variable definition was missing and causing an exception (#1456).
 
 ### Contrib
 - Check MK Statistics Cronjob: